fix các lỗi của nhanh fix: template component

+++ src/controllers/firmware.controller.ts
@@ -16,7 +16,10 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     createFirmware = async (req: Request, res: Response, next: NextFunction) => {
-        const employeeId = 'admin123';
+        const employeeId = req.user?.employeeId;
+        console.log("nhân viên", employeeId)
+        if (!employeeId) throwError(ErrorCodes.UNAUTHORIZED, 'Employee not authenticated');
+
         try {
             const firmware = await this.firmwareService.createFirmware(req.body, employeeId);
             res.status(201).json(firmware);
@@ -32,7 +35,8 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     updateFirmware = async (req: Request, res: Response, next: NextFunction) => {
-        const employeeId = 'admin123';
+        const employeeId = req.user?.employeeId;
+        if (!employeeId) throwError(ErrorCodes.UNAUTHORIZED, 'Employee not authenticated');
 
         try {
             const { firmwareId } = req.params;
@@ -50,14 +54,20 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     deleteFirmware = async (req: Request, res: Response, next: NextFunction) => {
-        const employeeId = 'admin123';
+        const accountId = req.user?.accountId;
+        
+        if (!accountId) {
+            return throwError(ErrorCodes.UNAUTHORIZED, 'Vui lòng đăng nhập để thực hiện hành động này');
+        }
+        
+        const employeeId = accountId;
 
         try {
             const { firmwareId } = req.params;
             const response = await this.firmwareService.deleteFirmware(firmwareId, employeeId);
             
             console.log('response', response)
-            res.status(204).json(response);
+            res.status(200).json(response);
         } catch (error) {
             next(error);
         }
@@ -70,11 +80,15 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     getFirmwareById = async (req: Request, res: Response, next: NextFunction) => {
+        const employeeId = req.user?.employeeId;
+        if (!employeeId) throwError(ErrorCodes.UNAUTHORIZED, 'Employee not authenticated');
+
         try {
             const { firmwareId } = req.params;
             const firmware = await this.firmwareService.getFirmwareById(firmwareId);
-            res.json(firmware);
+            res.status(200).json(firmware);
         } catch (error) {
+            console.log("err",error)
             next(error);
         }
     };
@@ -86,6 +100,12 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     getFirmwares = async (req: Request, res: Response, next: NextFunction) => {
+        const accountId = req.user?.accountId;
+        
+        if (!accountId) {
+            return throwError(ErrorCodes.UNAUTHORIZED, 'Vui lòng đăng nhập để thực hiện hành động này');
+        }
+        
         try {
             const firmwares = await this.firmwareService.getFirmwares();
             res.json(firmwares);
@@ -101,6 +121,12 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     getFirmwaresByTemplateId = async (req: Request, res: Response, next: NextFunction) => {
+        const accountId = req.user?.accountId;
+        
+        if (!accountId) {
+            return throwError(ErrorCodes.UNAUTHORIZED, 'Vui lòng đăng nhập để thực hiện hành động này');
+        }
+        
         try {
             const { templateId } = req.params;
             const firmwares = await this.firmwareService.getFirmwaresByTemplateId(templateId);
@@ -117,12 +143,19 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     confirmFirmwareByTester = async (req: Request, res: Response, next: NextFunction) => {
-        const employeeId = 'admin123';
+        const accountId = req.user?.accountId;
+        
+        if (!accountId) {
+            return throwError(ErrorCodes.UNAUTHORIZED, 'Vui lòng đăng nhập để thực hiện hành động này');
+        }
+        
+        const employeeId = accountId;
         try {
             const response = await this.firmwareService.confirmFirmwareByTester(req.body.firmwareId, employeeId, req.body.testResult);
             
             res.status(200).json(response);
         } catch (error) {
+            console.log(error)
             next(error);
         }
     };
@@ -134,7 +167,13 @@ class FirmwareController {
      * @param next Middleware tiếp theo
      */
     confirmFirmwareByRD = async (req: Request, res: Response, next: NextFunction) => {
-        const employeeId = 'admin123';
+        const accountId = req.user?.accountId;
+        
+        if (!accountId) {
+            return throwError(ErrorCodes.UNAUTHORIZED, 'Vui lòng đăng nhập để thực hiện hành động này');
+        }
+        
+        const employeeId = accountId;
         try {
             const response = await this.firmwareService.confirmFirmwareByRD(req.body.firmwareId, employeeId, req.body.testResult);
             
@@ -145,6 +184,11 @@ class FirmwareController {
     };
 
     getLatestVersionFirmwaresByTemplate = async (req: Request, res: Response, next: NextFunction) => {
+        const accountId = req.user?.accountId;
+
+        if (!accountId) {
+            return throwError(ErrorCodes.UNAUTHORIZED, 'Vui lòng đăng nhập để thực hiện hành động này');
+        }
         try {
             const latestVersionFirmwares = await this.firmwareService.getLatestVersionFirmwaresByTemplate();
             res.json(latestVersionFirmwares);
